requirements.txt
pyyaml
pytest
pytest-html
pytest-ordering
pytest-playwright
pytest-asyncio
